+++ src/hooks/index.js
@@ -315,8 +315,9 @@ exports.checkCorrectCourseOrTeamId = async (context) => {
 	if (courseGroupId || courseId) {
 		const userId = context.params.account.userId.toString();
 		// make it sense?
-		const validatedCourseId = (courseId || '').toString() || (context.id || '').toString();
+		let validatedCourseId = (courseId || '').toString() || (context.id || '').toString();
 		let query = {
+			_id: validatedCourseId,
 			teacherIds: {
 				$in: [userId],
 			},
@@ -325,7 +326,10 @@ exports.checkCorrectCourseOrTeamId = async (context) => {
 
 		if (courseGroupId) {
 			delete context.data.courseId;
+			const courseGroup = context.app.service('courseGroups').get(courseGroupId);
+			validatedCourseId = courseGroup.courseId;
 			query = {
+				_id: validatedCourseId,
 				$or: [
 					{ teacherIds: { $in: [userId] } },
 					{ userIds: { $in: [userId] } },
@@ -334,9 +338,9 @@ exports.checkCorrectCourseOrTeamId = async (context) => {
 			};
 		}
 
-		const course = await context.app.service('courses').get(validatedCourseId, { query });
+		const course = await context.app.service('courses').find({ query });
 
-		if (course === null) {
+		if (course.total !== 1) {
 			throw new Forbidden("The entered course doesn't belong to you!");
 		}
 		return context;
@@ -522,67 +526,6 @@ exports.mapPayload = (context) => {
 	});
 	return context;
 };
-
-exports.restrictToUsersOwnLessons = (context) => getUser(context).then((user) => {
-	if (testIfRoleNameExist(user, ['superhero', 'administrator'])) {
-		return context;
-	}
-	// before-hook
-	if (context.type === 'before') {
-		let populate = context.params.query.$populate;
-		if (typeof (populate) === 'undefined') {
-			populate = ['courseId', 'courseGroupId'];
-		} else if (Array.isArray(populate) && !populate.includes('courseId')) {
-			populate.push('courseId');
-			populate.push('courseGroupId');
-		}
-		context.params.query.$populate = populate;
-	} else {
-		// after-hook
-		if (context.method === 'get' && (context.result || {})._id) {
-			let tempLesson = [context.result];
-			tempLesson = tempLesson.filter((lesson) => {
-				if ('courseGroupId' in lesson) {
-					return userIsInThatCourse(user, lesson.courseGroupId, false);
-				}
-				return userIsInThatCourse(user, lesson.courseId, true)
-					|| (context.params.query.shareToken || {}) === (lesson.shareToken || {});
-			});
-			if (tempLesson.length === 0) {
-				throw new Forbidden("You don't have access to that lesson.");
-			}
-			if ('courseGroupId' in context.result) {
-				context.result.courseGroupId = context.result.courseGroupId._id;
-			} else {
-				context.result.courseId = context.result.courseId._id;
-			}
-		}
-
-		if (context.method === 'find' && ((context.result || {}).data || []).length > 0) {
-			context.result.data = context.result.data.filter((lesson) => {
-				if ('courseGroupId' in lesson) {
-					return userIsInThatCourse(user, lesson.courseGroupId, false);
-				}
-				return userIsInThatCourse(user, lesson.courseId, true)
-					|| (context.params.query.shareToken || {}) === (lesson.shareToken || {});
-			});
-
-			if (context.result.data.length === 0) {
-				throw new NotFound('There are no lessons that you have access to.');
-			} else {
-				context.result.total = context.result.data.length;
-			}
-			context.result.data.forEach((lesson) => {
-				if ('courseGroupId' in lesson) {
-					lesson.courseGroupId = lesson.courseGroupId._id;
-				} else {
-					lesson.courseId = lesson.courseId._id;
-				}
-			});
-		}
-	}
-	return context;
-});
 
 exports.restrictToUsersOwnClasses = (context) => getUser(context).then((user) => {
 	if (testIfRoleNameExist(user, ['superhero', 'administrator', 'teacher'])) {