+++ CHANGELOG.md
@@ -9,12 +9,19 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ## Unreleased
 
+
+## [24.4.0] - 2020-8-31
+
 ### Added
 - SC-5612 - Adding search feature to the admintables for nuxt-client.
 
 ### Fixed
 
 - SC-6122 - Edusharing preload thumbnails in parallel. Edusharing authentication stabilisation.
+
+## [24.3.3] - 2020-08-28
+
+- SC-6469: prevent admin access to lessons admins shouldnt have access to.
 
 ## [24.3.2] - 2020-08-26